#92 Add access verification to coprocessors 0-13

+++ src/core/arm1176jzf_s/isa/smlaxy.cpp
@@ -22,17 +22,17 @@ namespace zero_mate::arm1176jzf_s::isa
 
     std::uint32_t CSMLAxy::Get_Rd_Idx() const noexcept
     {
-        return static_cast<std::uint32_t>((m_value >> 16U) & 0b1111U);
+        return (m_value >> 16U) & 0b1111U;
     }
 
     std::uint32_t CSMLAxy::Get_Rn_Idx() const noexcept
     {
-        return static_cast<std::uint32_t>((m_value >> 12U) & 0b1111U);
+        return (m_value >> 12U) & 0b1111U;
     }
 
     std::uint32_t CSMLAxy::Get_Rs_Idx() const noexcept
     {
-        return static_cast<std::uint32_t>((m_value >> 8U) & 0b1111U);
+        return (m_value >> 8U) & 0b1111U;
     }
 
     std::uint32_t CSMLAxy::Get_Rm_Idx() const noexcept
